requirements.txt
discord
python-dotenv
requests
db-sqlite3